Moved tests from var_names.py to test_query_var_names.py
