+++ kep/test/test_query_var_names.py
@@ -1,0 +1,17 @@
+from kep.file_io.common import get_var_abbr, get_unit_abbr
+from kep.query.var_names import get_title, get_unit
+
+
+def test_get_var_abbr():
+    assert get_var_abbr('PROD_E_TWh') == 'PROD_E'
+
+def test_get_unit_abbr():
+    assert get_unit_abbr('PROD_E_TWh') == 'TWh'
+
+def test_get_title():
+    assert get_title('CONSTR_yoy') == 'Объем работ по виду деятельности "Строительство"'
+    assert get_title('I_bln_rub') == 'Инвестиции в основной капитал'
+    assert get_title('I_yoy') == 'Инвестиции в основной капитал'
+
+def test_get_unit():
+    assert get_unit('CONSTR_yoy') == 'в % к аналог. периоду предыдущего года'
